se implementan todos los servicios de la vista reporte gerenciales, se agrega funcionalidad de imprimir reporte y se realizan ajustes generales

+++ BikeStore.App/BikeStore.App.Persistencia/AppRepositorio/IRepositorioInventario.cs
@@ -13,6 +13,6 @@ namespace BikeStore.App.Persistencia
         int AddInventario(Inventario Inventario);
         int UpdateInventario(Inventario Inventario);
         int DeleteInventario(Inventario Inventario);
-        int GetTotalInversion();
+        long GetTotalInversion();
     } 
 }
